# swagger.yaml
openapi: 3.0.3
info:
  title: Fmarket API Documentation
  description: Documentation for the Fmarket API
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/
...
paths:
...
  /Store/create/{uid}:
      post:
        tags:
          - store
        summary: Create a store
        operationId: createStore
        parameters:
          - name: uid
            in: path
            required: true
            schema:
              type: string
            description: User ID
        requestBody:
          description: Store data to create
          content:
            multipart/form-data:
              schema:
                type: object
                properties:
                  name:
                    type: string
                  image:
                    type: string
                  address:
                    type: string
        responses:
          '201':
            description: Store created successfully
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/SuccessMessage'
          '404':
            description: Account not found
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Error'
          '500':
            description: Internal server error
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Error'
  /Store/info/{storeId}:
      get:
        tags:
          - store
        summary: Get store details
        operationId: detailStore
        parameters:
          - name: storeId
            in: path
            required: true
            schema:
              type: string
            description: Store ID
        responses:
          '200':
            description: Store details retrieved successfully
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/StoreDetails'
          '404':
            description: Store not found
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Error'
          '500':
            description: Internal server error
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Error'
  /Store/delete/{storeId}:
    delete:
      tags:
        - store
      summary: Delete a store
      operationId: deleteStore
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: string
          description: Store ID
      responses:
        '200':
          description: Store deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '404':
          description: Store not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    SuccessMessage:
      type: object
      properties:
        code:
          type: number
          example: 200
        message:
          type: string
          example: "Success Message"
    Error:
      type: object
      properties:
        code:
          type: number
          example: 404
        message:
          type: string
          example: "Error Message"
    StoreDetails:
      type: object
      properties:
        code:
          type: number
          example: 200
        data:
          $ref: '#/components/schemas/StoreData'
        message:
          type: string
          example: "Get store success"
    StoreData:
      type: object
      properties:
        _id:
          type: string
          example: "12345"  # Replace with actual store ID
        name:
          type: string
          example: "Store Name"
        image:
          type: string
          example: "store_image.jpg"
        address:
          type: string
          example: "123 Store St, City"
        user_id:
          type: string
          example: "adkhfa"
        is_active:
          type: boolean
          example: true
...